Split trace export tests
